APYTHON/requirements.txt
fastapi
uvicorn
pandas
sqlalchemy
pymysql
python-dotenv
openpyxl